% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-pallettes.R
\name{color_palette}
\alias{color_palette}
\title{Color Pallettes}
\usage{
color_palette(n, type = "cat", palette = "Nelson Nygaard")
}
\arguments{
\item{n}{Number of Colors neeeded}

\item{type}{Type of color palette ("cat","seq","div")}

\item{palette}{Name of palette}
}
\value{
Vector of colors
}
\description{
Color Pallettes
}

