% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleMapUtils.R
\name{decodeLineR}
\alias{decodeLineR}
\title{Google Polyline Decoder}
\usage{
decodeLineR(encoded)
}
\arguments{
\item{encoded}{Encoded polyline result from Google Directions API request}
}
\value{
Decoded polyline.
}
\description{
Google Polyline Decoder
}
\references{
Function adapted from http://facstaff.unca.edu/mcmcclur/GoogleMaps/EncodePolyline/decode.js and https://gist.github.com/diegovalle/916889/895dba68c0f9f5398c9d9c75856126e233b9acd7
}

