% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-utilities.R
\name{expand_id_col}
\alias{expand_id_col}
\title{Expand ID Column Vector}
\usage{
expand_id_col(id_vec, n_rep)
}
\arguments{
\item{id_vec}{The vector of unique IDs you would like to 'expand' (i.e. turn [1,2] into [1,1,2,2])}

\item{n_rep}{The number of repetitions in the 'expansion'}
}
\value{
Expanded vector
}
\description{
Expand ID Column Vector
}

