% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gtfs-tools-pdx-bblanc.R
\name{exportRouteShape}
\alias{exportRouteShape}
\title{Export GTFS routes to ESRI shapefile}
\usage{
exportRouteShape(feedPath, outPath, shapeName, returnShape = FALSE)
}
\arguments{
\item{feedPath}{The relative or absolute path to the folder containing the GTFS feed text files}

\item{outPath}{The relative or absolute path to the folder where you would like to save the reuslting shapefile}

\item{shapeName}{The filename for your resulting shapefile}

\item{returnShape}{Boolean indicating if route shape should be returned}
}
\value{
Returns route shape as SpatialLinesDataFrame if desired
}
\description{
Export GTFS routes to ESRI shapefile
}

