% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gtfs-tools-pdx-bblanc.R, R/gtfs-tools.R
\name{fetchFeed}
\alias{fetchFeed}
\title{Fetch GTFS feed files}
\usage{
fetchFeed(feedName)

fetchFeed(feedName)
}
\arguments{
\item{feedName}{Name of transit agency (will try to find similar names if not exact match)}

\item{feedName}{Name of transit agency (will try to find similar names if not exact match)}
}
\value{
None

None
}
\description{
This function will match the requested transit agency name with a feed via the GTFS data exchange, and the latest feed files will be downloaded to 'feeds' subfolder of working directory.

This function will match the requested transit agency name with a feed via the GTFS data exchange, and the latest feed files will be downloaded to 'feeds' subfolder of working directory.
}

