% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleMapUtils.R
\name{gDirsToShape}
\alias{gDirsToShape}
\title{Convert Google directions to Spatial Object}
\usage{
gDirsToShape(googleDirs, mode)
}
\arguments{
\item{googleDirs}{List of resuls from `getGoogleDirections()` function}

\item{mode}{Transportation mode of directions ('driving','walking','bicycling','transit')}
}
\value{
SpatialLinesDataFrame object representing Google directed routes
}
\description{
Convert Google directions to Spatial Object
}

