% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-utilities.R
\name{geo_join}
\alias{geo_join}
\title{Join Census Geography to ACS Results}
\usage{
geo_join(table, geo_level)
}
\arguments{
\item{table}{Data frame of ACS results, such as the output of `transit_pops()`}

\item{geo_level}{A geographic level: county, place, tract, or block group}
}
\value{
SpatialPolygonsDataFrame with ACS data joined
}
\description{
Join Census Geography to ACS Results
}

