% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-utilities.R
\name{geo_join_lodes}
\alias{geo_join_lodes}
\title{Join aggregated lodes data to geography}
\usage{
geo_join_lodes(agg_table, type, geo_level, crosswalk, acs_geom = NULL)
}
\arguments{
\item{agg_table}{Aggregated LODES data frame (result of `lodes_agg()`)}

\item{geo_level}{Geographic level}

\item{crosswalk}{Crosswalk}

\item{acs_geom}{ACS geometry from ACS fetching/joining process}
}
\value{
SpatialPolygonsDataFrame with LODES data joined
}
\description{
Join aggregated lodes data to geography
}

