% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleMapUtils.R
\name{geocode_place}
\alias{geocode_place}
\title{Geocode a place name using Google Places API}
\usage{
geocode_place(placeString, key, output = "loc")
}
\arguments{
\item{placeString}{A string describing the place you'd like to geocode}

\item{output}{The output can either be a simple location ('loc') or a list with all response results ('all')}
}
\value{
Location or list of API results
}
\description{
Geocode a place name using Google Places API
}

