% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfgm_carparks.R
\name{get_tfgm_carparks}
\alias{get_tfgm_carparks}
\title{Get TfGM carparks Function}
\usage{
get_tfgm_carparks(readpath = getwd())
}
\arguments{
\item{readpath}{the path where you saved the file with the api key, defaults to working directory}
}
\description{
This function returns a dataframe of carparks in Greater Manchester
This dataset contains car park names, state (giving an indication of whether it is open or closed), capacity and real-time occupancy for selected car parks in the districts of Manchester, Stockport and Wigan only.
Metrolink Park & Ride sites will be available shortly.
Function also exracts longitude and latitude into separate columns from WKT column
}
\examples{
carpark_df <- get_tfgm_carparks()
}
\keyword{api,}
\keyword{carparks}
\keyword{tfgm,}
\keyword{transport,}
\keyword{uk,}
