% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleMapUtils.R
\name{getGoogleDirections}
\alias{getGoogleDirections}
\title{Fetch Google directions from Directions API}
\usage{
getGoogleDirections(from, to, key, mode, departure = NULL, arrival = NULL,
  alternatives = FALSE, waypoints = NULL)
}
\arguments{
\item{from}{Start location to fetch directions for (vector in [Lat Lng] format)}

\item{to}{End location to fetch directions for (vector in [Lat Lng] format)}

\item{key}{API key for Google Directions API (from Google Developers Console)}

\item{mode}{Travel mode: one of four strings: 'driving', 'walking', 'bicycling', 'transit'}

\item{departure}{Optional departure time of POSIXct class. Cannot specify both arrival and departure times.}

\item{arrival}{Optional arrival time of POSIXct class. Cannot specify both arrival and departure times.}

\item{waypoints}{Optional additional locations to travel through along route. A data frame or with 'lat' and 'lng' columns, and an observation for each destination.}
}
\value{
Encoded google directions string
}
\description{
Fetch Google directions from Directions API
}

