% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gtfs-tools.R
\name{gtfs_to_excel}
\alias{gtfs_to_excel}
\title{Import GTFS tables into a single Excel file}
\usage{
gtfs_to_excel(feedPath)
}
\arguments{
\item{feedPath}{A relative or absolute file path to the feed you want to import into Excel}
}
\value{
Writes Excel file to directory of feed
}
\description{
Import GTFS tables into a single Excel file
}

