% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial-utilities.R
\name{IDs}
\alias{IDs}
\alias{IDs.SpatialPolygonsDataFrame}
\alias{IDs.default}
\title{Get sp feature IDs}
\usage{
IDs(x, ...)

\method{IDs}{default}(x, ...)

\method{IDs}{SpatialPolygonsDataFrame}(x, ...)
}
\arguments{
\item{x}{The object to get the IDs from}

\item{\dots}{Pass-alongs}
}
\description{
Get sp feature IDs
}

