% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general-utilities.R
\name{list2frame}
\alias{list2frame}
\title{Converts Two-Level List to Data Frame}
\usage{
list2frame(list)
}
\arguments{
\item{list}{List with One level of sub-lists (sub-lists cannot contain lists)}
}
\value{
Data frame
}
\description{
Converts Two-Level List to Data Frame
}

