% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-utilities.R
\name{lodes_agg}
\alias{lodes_agg}
\title{Aggregate LODES data to higher geography}
\usage{
lodes_agg(table, type, geo_level, crosswalk)
}
\arguments{
\item{table}{LODES data frame (result of `lodes_fetch()`)}

\item{type}{Character string indicating what LODES file type you are aggregating ('OD','WAC', or 'RAC')}

\item{geo_level}{Character string indicating what geographic level to aggregate to (block group, tract, place, or county)}

\item{crosswalk}{Crosswalk data frame (result of `lodes_crosswalk()`)}
}
\value{
Data frame of aggregated LODES data
}
\description{
Aggregate LODES data to higher geography
}

