% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-utilities.R
\name{lodes_crosswalk}
\alias{lodes_crosswalk}
\title{Fetch LODES data crosswalk for given state}
\usage{
lodes_crosswalk(state_code)
}
\arguments{
\item{state_code}{Two character state code (string)}
}
\value{
Data frame containing LODES crosswalk
}
\description{
Fetch LODES data crosswalk for given state
}

