% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-utilities.R
\name{lodes_fetch}
\alias{lodes_fetch}
\title{Fetch LODES data files for given state and year (OD, RAC, or WAC)}
\usage{
lodes_fetch(state_code, file_type, year)
}
\arguments{
\item{state_code}{Two character state code (string)}

\item{file_type}{Character string indicating what file type you want to fetch ('OD','WAC', or 'RAC')}

\item{year}{Year of interest}
}
\value{
Data frame containing all LODES data requested
}
\description{
Fetch LODES data files for given state and year (OD, RAC, or WAC)
}

