% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general-utilities.R
\name{naturalBreaks}
\alias{naturalBreaks}
\title{Calculate natural breaks for specified fields in table}
\usage{
naturalBreaks(table, fields, numBreaks)
}
\arguments{
\item{table}{The data frame or matrix object with the data and fields you need to calculate breaks for}

\item{fields}{A vector of character strings corresponding to the fields you need to calculate breaks for}

\item{numBreaks}{The number of break points to calculate}
}
\value{
Data frame with break points for each field specified
}
\description{
This function will calculate natural (Jenks) breaks for each specified field in the input table and return a table of the break
points based on the number of breaks desired.
}

