% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial-utilities.R
\name{rbind.SpatialPolygonsDataFrame}
\alias{rbind.SpatialPolygonsDataFrame}
\title{rbind SpatialPolygonsDataFrames together, fixing IDs if duplicated}
\usage{
\method{rbind.SpatialPolygonsDataFrame}{}(..., fix.duplicated.IDs = TRUE)
}
\arguments{
\item{fix.duplicated.IDs}{Whether to de-duplicate polygon IDs or not}

\item{\dots}{SpatialPolygonsDataFrame(s) to rbind together}
}
\value{
SpatialPolygonsDataFrame
}
\description{
rbind SpatialPolygonsDataFrames together, fixing IDs if duplicated
}
\author{
Ari B. Friedman, with key functionality by csfowler on StackExchange
}

