% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleMapUtils.R
\name{route2shape}
\alias{route2shape}
\title{Convert decoded Google Directions polyline into SpatialLines object}
\usage{
route2shape(directions)
}
\arguments{
\item{directions}{Decoded polyline directions from Google Directions API}
}
\value{
SpatialLines object representing route received from Google Directions API
}
\description{
Convert decoded Google Directions polyline into SpatialLines object
}

