% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_key.R
\name{save_key}
\alias{save_key}
\title{Save key Function}
\usage{
save_key(org, app_key, app_id = NULL, writepath = getwd())
}
\arguments{
\item{org}{code of transport authority the key is for (tfl for Transport for London, tfgm for Transport for Greater Manchester)}

\item{app_key}{the API key for your application}

\item{app_id}{the API id for your application (TfL only)}

\item{writepath}{the path where you want to save the file, defaults to working directory}
}
\description{
This function allows you to save your api key to a .txt file, which you can read it from in the other functions.
}
\examples{
save_key("tfl", "0f2fd687", "6c23ac1o45cfec9hhd6bf32b393a84do3") #these are not real api key/id values
}
\keyword{api,}
\keyword{transport,}
\keyword{uk}
