% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gtfs-tools-pdx-bblanc.R, R/gtfs-tools.R
\name{spatialTransitShapes}
\alias{spatialTransitShapes}
\title{Create SpatialLines from GTFS shapes.txt file}
\usage{
spatialTransitShapes(feedPath)

spatialTransitShapes(feedPath)
}
\arguments{
\item{feedPath}{Absolute or relative path to the folder containing GTFS feed files.}

\item{feedPath}{Absolute or relative path to the folder containing GTFS feed files.}
}
\value{
SpatialLines object corresponding to shapes.txt file

SpatialLines object corresponding to shapes.txt file
}
\description{
Create SpatialLines from GTFS shapes.txt file

Create SpatialLines from GTFS shapes.txt file
}

