% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/acs-utilities.R
\name{transit_pops}
\alias{transit_pops}
\title{Fetch transit dependent population statistics}
\usage{
transit_pops(geog, endyear = 2014, span = 5)
}
\arguments{
\item{geog}{Geography object (constructed using `geo.make()` from `acs` package)}

\item{endyear}{Endyear of ACS data request (default is 2014)}

\item{span}{Span of ACS estimate (default is five years)}
}
\value{
Dataframe with transit dependent population statistics.
}
\description{
Requires dplyr,tidyr, and acs packages to be installed.
}

