% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleMapUtils.R
\name{travel_time}
\alias{travel_time}
\title{Query travel times from Google API}
\usage{
travel_time(from, to, key, mode, departure = NULL, arrival = NULL,
  tt_type = "fastest")
}
\arguments{
\item{from}{Start location to fetch directions for (vector in [Lat Lng] format)}

\item{to}{End location to fetch directions for (vector in [Lat Lng] format)}

\item{key}{API key for Google Directions API (from Google Developers Console)}

\item{mode}{Travel mode: one of four strings: 'driving', 'walking', 'biking', 'transit'}

\item{departure}{Optional departure time of POSIXct class. Cannot specify both arrival and departure times.}

\item{arrival}{Optional arrival time of POSIXct class. Cannot specify both arrival and departure times.}
}
\value{
Travel time (minutes)
}
\description{
Query travel times from Google API
}

