% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleMapUtils.R
\name{tt_matrix}
\alias{tt_matrix}
\title{Title}
\usage{
tt_matrix(origins, destinations, places = FALSE, key, mode,
  departure = NULL, arrival = NULL)
}
\arguments{
\item{origins}{Data frame with origin data with column names 'id','lat','lng' or 'id','place' (if places=TRUE)}

\item{destinations}{Data frame with destination data with column names 'id','lat','lng' or 'id','place' (if places=TRUE)}

\item{places}{Boolean indicating if you are providing lat/lng locations or place names}

\item{key}{Google Maps API Key}

\item{mode}{Travel mode}

\item{departure}{POSIXct departure time}

\item{arrival}{POSIXct arrival time}
}
\value{
Data frame with travel times and distances identified by origin and destination id
}
\description{
Title
}

