% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleMapUtils.R
\name{getGoogleDirections}
\alias{getGoogleDirections}
\title{Fetch Google directions from Directions API}
\usage{
getGoogleDirections(from, to, key)
}
\arguments{
\item{from}{Start location to fetch directions for (vector in [Lat Lng] format)}

\item{to}{End location to fetch directions for (vector in [Lat Lng] format)}

\item{key}{API key for Google Directions API (from Google Developers Console)}
}
\value{
Encoded google directions string
}
\description{
Fetch Google directions from Directions API
}
\references{
Adapted from function here: https://gist.github.com/diegovalle/916889/895dba68c0f9f5398c9d9c75856126e233b9acd7
}

