% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleMapUtils.R
\name{lines2points}
\alias{lines2points}
\title{Convert SpatialLines to coordinate data frame for map matching.}
\usage{
lines2points(linesShape, trip_id)
}
\arguments{
\item{linesShape}{SpatialLines shape containing trip to convert to points}

\item{trip_id}{ID of trip to convert to points}
}
\value{
Coordinate data frame to save as .csv for map matching.
}
\description{
Convert SpatialLines to coordinate data frame for map matching.
}

